@@ -3,30 +3,10 @@ description: Waatea 2 Client App.
 
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
-version: 0.0.39+1
+version: 0.0.41+1
 
 environment:
-  sdk: '>=2.12.0 <4.0.0'
-
-fonts:
-  - family: KaTeX
-    fonts:
-      - asset: packages/flutter_math_fork/fonts/KaTeX_Main-Regular.ttf
-      - asset: packages/flutter_math_fork/fonts/KaTeX_Main-Bold.ttf
-      - asset: packages/flutter_math_fork/fonts/KaTeX_Main-Italic.ttf
-      - asset: packages/flutter_math_fork/fonts/KaTeX_Main-BoldItalic.ttf
-      - asset: packages/flutter_math_fork/fonts/KaTeX_Math-Italic.ttf
-      - asset: packages/flutter_math_fork/fonts/KaTeX_SansSerif-Regular.ttf
-      - asset: packages/flutter_math_fork/fonts/KaTeX_SansSerif-Bold.ttf
-      - asset: packages/flutter_math_fork/fonts/KaTeX_SansSerif-Italic.ttf
-      - asset: packages/flutter_math_fork/fonts/KaTeX_SansSerif-BoldItalic.ttf
-      - asset: packages/flutter_math_fork/fonts/KaTeX_Typewriter-Regular.ttf
-      - asset: packages/flutter_math_fork/fonts/KaTeX_Typewriter-Bold.ttf
-      - asset: packages/flutter_math_fork/fonts/KaTeX_Fraktur-Regular.ttf
-      - asset: packages/flutter_math_fork/fonts/KaTeX_Fraktur-Bold.ttf
-      - asset: packages/flutter_math_fork/fonts/KaTeX_Fraktur-Italic.ttf
-      - asset: packages/flutter_math_fork/fonts/KaTeX_Fraktur-BoldItalic.ttf
-
+  sdk: '>=2.17.0 <4.0.0'
 
 dependencies:
   flutter:
@@ -36,35 +16,45 @@ dependencies:
   provider: ^6.0.0
   graphql_flutter: ^5.1.2
   shared_preferences: ^2.2.0
-  intl: ^0.18.1
+  intl: ^0.19.0
   random_avatar: any
   url_launcher: ^6.1.12
   whatsapp_unilink: ^2.1.0
-  csv: ^5.0.2
-  file_picker: ^6.1.1
+  csv: ^6.0.0
+  file_picker: ^9.2.1
   file_saver: ^0.2.6
   file_selector: ^1.0.0
   universal_html: ^2.2.3
-  package_info_plus: ^4.1.0
-  flutter_quill: ^7.4.2
-  flutter_quill_extensions: ^0.5.0
+  package_info_plus: ^8.3.0
+  flutter_quill: ^11.1.2
+  flutter_quill_extensions: ^11.0.0
   flutter_reorderable_list: ^1.3.1
   pdf: ^3.10.4
   open_file: ^3.3.2
   flutter_svg: ^2.0.7
-  screenshot: ^2.1.0
-  share_plus: ^7.2.1
+  screenshot: ^3.0.0
+  share_plus: ^10.1.4
   path_provider: ^2.1.1
   multi_select_flutter: ^4.1.3
+  permission_handler: ^12.0.0+1
+  gallery_saver: ^2.3.2
+
+dependency_overrides:
+  archive: ^4.0.5
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^3.0.1
+  flutter_lints: ^5.0.0
 
 flutter:
   uses-material-design: true
+
+  fonts:
+    - family: Gilroy
+      fonts:
+        - asset: assets/fonts/Gilroy-Black.ttf
 
   assets:
     - assets/shuttle_run.png
@@ -72,3 +62,4 @@ flutter:
     - assets/scotland_drill.png
     - assets/complex_shuttles.png
     - assets/broken_broncos.png
+    - assets/template.png
